+++ components/index.ts
@@ -1,9 +1,21 @@
 import QRCode, { props } from './qrcode/index'
 
 import type { App } from 'vue'
-import type { QRCodeInst, QRCodeLevel, QRCodeStatus } from './qrcode/src/types'
+import type {
+  QRCodeInst,
+  QRCodeLevel,
+  QRCodeStatus,
+  QRCodeComponentProps,
+  QRCodeOptions,
+} from './qrcode/src/types'
 
-export type { QRCodeInst, QRCodeLevel, QRCodeStatus }
+export type {
+  QRCodeInst,
+  QRCodeLevel,
+  QRCodeStatus,
+  QRCodeComponentProps,
+  QRCodeOptions,
+}
 
 export { QRCode as Vue3NextQrcode, props as qrcodeProps }
 export default {