+++ components/qrcode/src/props.ts
@@ -1,6 +1,6 @@
 import type { MaybeArray, QRCodeStatus, QRCodeLevel } from './types'
 import type { PropType, VNode } from 'vue'
-import type { ComponentOptions } from 'awesome-qr/lib/awesome-qr'
+import type { ComponentOptions } from '../core/awesome-qr'
 
 const props = {
   maskColor: {